refactor: improve component transformer namings
