+++ lib/mdxish.ts
@@ -26,9 +26,9 @@ import codeTabsTransformer from '../processor/transform/code-tabs';
 import embedTransformer from '../processor/transform/embeds';
 import imageTransformer from '../processor/transform/images';
 import { closeSelfClosingHtmlTags } from '../processor/transform/mdxish/close-self-closing-html-tags';
-import mdxishComponentBlocks from '../processor/transform/mdxish/components/component-blocks';
-import mdxishInlineComponentBlocks from '../processor/transform/mdxish/components/inline-component-blocks';
-import mdxishInlineComponents from '../processor/transform/mdxish/components/inline-components';
+import mdxishInlineComponentBlocks from '../processor/transform/mdxish/components/inline-html';
+import mdxishInlineComponents from '../processor/transform/mdxish/components/inline-mdx-blocks';
+import mdxishComponentBlocks from '../processor/transform/mdxish/components/mdx-blocks';
 import mdxishSelfClosingBlocks from '../processor/transform/mdxish/components/self-closing-blocks';
 import { processSnakeCaseComponent } from '../processor/transform/mdxish/components/snake-case-components';
 import evaluateExpressions from '../processor/transform/mdxish/evaluate-expressions';
